+++ R/Class-SimParam.R
@@ -2009,7 +2009,10 @@ SimParam = R6Class(
     #' SP$setRecombRatio(2) #Twice as much recombination in females
     setRecombRatio = function(femaleRatio){
       stopifnot(femaleRatio>0)
+      # Both must be read before .sepMap is set, because the accessors
+      # switch to averaging the sex-specific values once it is TRUE
       genMap = self$genMap
+      centromere = self$centromere
       private$.sepMap = TRUE
       feSc = 2/(1/femaleRatio+1)
       maSc = 2/(femaleRatio+1)
@@ -2017,12 +2020,12 @@ SimParam = R6Class(
                                   function(x){
                                     feSc*x
                                   })
-      private$.femaleCentromere = feSc*private$.femaleCentromere
+      private$.femaleCentromere = feSc*centromere
       private$.maleMap = lapply(genMap,
                                 function(x){
                                   maSc*x
                                 })
-      private$.maleCentromere = maSc*private$.maleCentromere
+      private$.maleCentromere = maSc*centromere
       invisible(self)
     },